Remove unused variable f in temp.from.geom
In section 1.7, in the definition of the temp.from.geom function, there is the variable f defined, I guess, to avoid the ambiguity of sign in case fi is 0, but then not used. It is not really impacting but a bit confusing, and since sign(0) return 0 I think it is even more reasonable to show a perfectly flat curve.