Mismatch between the values shown/computed
I noticed that the result of summary(m@spModel$learner.model$super.model$learner.model) in GitHub does not match the one described (that instead match with the one obtained). Maybe the cause of this should be investigated, or in any case write a more conservative sentence that does not mention precise values.